requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile
#
ed25519==1.5
    # via nkeys
nats-py[nkeys]==2.7.2
    # via -r requirements.in
nkeys==0.1.0
    # via nats-py
